+++ crates/toolchain-plugin/src/toolchain_plugin.rs
@@ -3,7 +3,7 @@ use moon_feature_flags::glob_walk;
 use moon_pdk_api::*;
 use moon_plugin::{Plugin, PluginContainer, PluginId, PluginRegistration, PluginType};
 use proto_core::flow::install::InstallOptions;
-use proto_core::{PluginLocator, Tool, ToolSpec, UnresolvedVersionSpec};
+use proto_core::{PluginLocator, Tool, ToolSpec, UnresolvedVersionSpec, locate_tool};
 use starbase_utils::glob::GlobSet;
 use std::fmt;
 use std::ops::Deref;
@@ -399,35 +399,49 @@ impl ToolchainPlugin {
     ) -> miette::Result<SetupToolchainOutput> {
         let mut output = SetupToolchainOutput::default();
 
-        // Only install if a version has been configured,
-        // and the plugin provides the required APIs
-        if let (Some(version), Some(tool)) = (&input.configured_version, &self.tool) {
+        if let Some(tool) = &self.tool {
             let mut tool = tool.write().await;
-            let spec = ToolSpec::new(version.to_owned());
 
-            // Resolve the version first so that it is available
-            input.version = Some(tool.resolve_version(&spec, false).await?);
+            // Only install if a version has been configured
+            if let Some(version) = &input.configured_version {
+                let spec = ToolSpec::new(version.to_owned());
 
-            // Only setup if not already been
-            if !tool.is_setup(&spec).await? {
-                on_setup()?;
+                // Resolve the version first so that it is available
+                input.version = Some(tool.resolve_version(&spec, false).await?);
 
-                output.installed = tool
-                    .setup(
-                        &spec,
-                        InstallOptions {
-                            skip_prompts: true,
-                            skip_ui: true,
-                            ..Default::default()
-                        },
-                    )
-                    .await?
-                    .is_some();
+                // Only setup if not already been
+                if !tool.is_setup(&spec).await? {
+                    on_setup()?;
+
+                    output.installed = tool
+                        .setup(
+                            &spec,
+                            InstallOptions {
+                                skip_prompts: true,
+                                skip_ui: true,
+                                ..Default::default()
+                            },
+                        )
+                        .await?
+                        .is_some();
+                }
+
+                // Locate pieces that we'll need
+                tool.locate_exes_dirs().await?;
+                tool.locate_globals_dirs().await?;
             }
 
-            // Locate pieces that we'll need
-            tool.locate_exes_dirs().await?;
-            tool.locate_globals_dirs().await?;
+            // Pre-load the tool plugin so that task executions
+            // avoid network race conditions and collisions
+            if let Ok(loader) = tool.proto.get_plugin_loader() {
+                if let Some(locator) = tool
+                    .locator
+                    .clone()
+                    .or_else(|| locate_tool(&tool.id, &tool.proto).ok())
+                {
+                    let _ = loader.load_plugin(&tool.id, &locator).await;
+                }
+            }
         }
 
         // This should always run, regardless of the install outcome